Adding more codes

+++ workspace-dadlabs/lab03/src/lab03/Exercise5Reader.java
@@ -8,16 +8,13 @@ public class Exercise5Reader {
 	public static void main(String[] args) {
 
 	
-	// 1. Declare output file 
 			String sourceFile = "Exercise5.txt";
 			System.out.println("Reading data from " + sourceFile + "\n");
 
 			try {
 
-				// 2. Create stream to read data
 				BufferedInputStream bis = new BufferedInputStream(new FileInputStream(sourceFile));
 
-				// Variables for processing byte-based data
 				double Oct16 = 0;
 				double Oct17 = 0;
 				double Oct18 = 0;
@@ -25,13 +22,11 @@ public class Exercise5Reader {
 				double Oct20 = 0;
 				double Oct21 = 0;
 
-				double totalUtilization = 0;
+				int totalUtilization = 0;
 				int noOfRecords = 0;
 
-				// 3. Process data until end-of-file
 				while(bis.available() > 0) {
 
-					// Read data
 					Oct16 = bis.read();
 					Oct17 = bis.read();
 					Oct18 = bis.read();
@@ -50,10 +45,8 @@ public class Exercise5Reader {
 					noOfRecords ++;
 				}
 
-				// 4. Close stream
 				bis.close();
 
-				// Calculate average utilization
 				double averageUtlization = totalUtilization / noOfRecords;
 				System.out.print("\nAverage Daily Rainfall for" + " 6 stations: " + averageUtlization + "%");
 
@@ -62,7 +55,6 @@ public class Exercise5Reader {
 				ex.printStackTrace();
 			}
 
-			// Indicate end of program - Could be successful
 			System.out.println("\nEnd of program.");
 
 		}
